// src/api/generated/models/getCompaniesCompanyIdInvoicesDirection.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type GetCompaniesCompanyIdInvoicesDirection = typeof GetCompaniesCompanyIdInvoicesDirection[keyof typeof GetCompaniesCompanyIdInvoicesDirection];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetCompaniesCompanyIdInvoicesDirection = {
  inbound: 'inbound',
  outbound: 'outbound',
} as const;