// src/api/generated/models/supplierInvoiceSummaryStatus.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type SupplierInvoiceSummaryStatus = typeof SupplierInvoiceSummaryStatus[keyof typeof SupplierInvoiceSummaryStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SupplierInvoiceSummaryStatus = {
  DRAFT: 'DRAFT',
  ISSUED: 'ISSUED',
  PARTIALLY_PAID: 'PARTIALLY_PAID',
  PAID: 'PAID',
  CONFIRMED: 'CONFIRMED',
} as const;